BACKEND-27 BACKEND-28 BACKEND-29 BACKEND-30
TASK 27 - 30 were completed

I created end points for getting appts based on role type.

End points were also created for updating appts and also changing the status of an appt. Customers should only be able to cancel appts as its set to default when made and only owners, barbers, or admins be able to confirm if they made the appt or where a no show.

Created a function in AppointmentServices that checks whether or not there are any potential overlaps with other appts. If so will return that appt cannot be made.