Fixed Schedule APIs
Made it so that POST and PATCH for schedule api's barber_id is obtained rather than given through payload. The GET function will still need the barber_id given so that it's versatile

Also made it so barbers are the only one able to set their availability that includes updating and setting it.